Dockerfile installs Swift kernel successfully
This commit resolves permissions error due to iSwiftKernel
being created and compiled as root in the Docker image
build process.

Now, it is still created and built as root, but it is
chowned to the notebook user (jovyan, aka NB_USER) and
the kernelspec install process is run as the notebook
user.

This makes it possible to run the docker image, create a
new notebook, and select Swift as the language. But there's
still some kind of problem with the kernel itself that
prevents it from responding.